Make the figure a widget
